+++ CHANGELOG.md
@@ -1,11 +1,26 @@
 ## MapLibre GL widget v0.x.x
 
+- Add control position setting
+- Add fullscreen control option
+- Add a button for setting the current pitch as the **Initial Pitch** setting. 
+  This button is only available when using WireCloud v1.4 in edit mode.
+- Add a button for setting the current zoom level as the **Initial Zoom
+    Level** setting. This button is only available when using WireCloud v1.4 in
+    edit mode. (Porting of https://github.com/Wirecloud/ol3-map-widget/)
+- Add a button for setting both the **Initial Location** and the **Initial
+    Zoom Level** settings from the current displayed area. This button is only
+    available when using WireCloud v1.4 in edit mode.
+    (Porting of https://github.com/Wirecloud/ol3-map-widget/)
+- Add a button to set the **Initial Location** setting using the center of the
+    current displayed area. This button is only available when using WireCloud
+    v1.4 in edit mode.
+    (Porting of https://github.com/Wirecloud/ol3-map-widget/)
 - Add Visible PoIs feature
-- Fix PoI selected
+- Fix PoI selected (#5)
 
 ## MapLibre GL widget v0.2.0 - 2021/4/11
 
-- Add catalogue icon
+- Add catalogue icon (#4)
 - Fix removePoI error (#3)
 - Improve preference settings (#3)
 - Improve rotateCamera command (#3)