+++ src/doc/changelog.md
@@ -1,5 +1,18 @@
 ## v0.x.x
 
+- Add control position setting
+- Add fullscreen control option
+- Add a button for setting the current pitch as the **Initial Pitch** setting.
+    This button is only available when using WireCloud v1.4 in edit mode.
+- Add a button for setting the current zoom level as the **Initial Zoom
+    Level** setting. This button is only available when using WireCloud v1.4 in
+    edit mode.
+- Add a button for setting both the **Initial Location** and the **Initial
+    Zoom Level** settings from the current displayed area. This button is only
+    available when using WireCloud v1.4 in edit mode.
+- Add a button to set the **Initial Location** setting using the center of the
+    current displayed area. This button is only available when using WireCloud
+    v1.4 in edit mode.
 - Add Visible PoIs feature
 - Fix PoI selected